AUTh cjanges done

+++ routers/posts.py
@@ -1,36 +1,45 @@
 from typing import Annotated
 
-from fastapi import FastAPI, APIRouter, Request, HTTPException, status, Path, Depends
-from sqlalchemy import select, func
+from fastapi import APIRouter, HTTPException, status, Depends
+from sqlalchemy import select
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.orm import selectinload
 
 import models
+from auth import CurrentUser
 from databsase_async import Base, engine, get_db
-from schemas import PostCreate, PostResponse, PostUpdate, UserCreate, UserResponse, UserUpdate
+from schemas import PostCreate, PostResponse, PostUpdate, UserCreate, UserPublic, UserUpdate
 
 router = APIRouter()
 
 @router.get('', response_model=list[PostResponse])
-async def get_posts(db: Annotated[AsyncSession, Depends(get_db)]):
+async def get_posts(current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
+  if not current_user.id:
+    raise HTTPException(
+        status_code=status.HTTP_403_FORBIDDEN,
+        detail="Not authorized to get the post details"
+      )
   result = await db.execute(select(models.Post).options(selectinload(models.Post.author)))
   posts = result.scalars().all()
   return posts
 
 
 @router.post("", response_model=PostResponse, status_code=status.HTTP_201_CREATED)
-async def create_post(post: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
-   result = await db.execute(select(models.User).where(models.User.id == post.user_id))
-   user = result.scalars().first()
-   if not user:
-     raise HTTPException(
-       status_code=status.HTTP_404_NOT_FOUND,
-       detail="USer not found"
-     )
+async def create_post(post: PostCreate,
+current_user:CurrentUser,
+ db: Annotated[AsyncSession, Depends(get_db)]):
+  #  result = await db.execute(select(models.User).where(models.User.id == post.user_id))
+  #  user = result.scalars().first()
+  #  if not user:
+  #    raise HTTPException(
+  #      status_code=status.HTTP_404_NOT_FOUND,
+  #      detail="USer not found"
+  #    )
+
    new_post = models.Post(
      title = post.title,
      content = post.title,
-     user_id = post.user_id
+     user_id = current_user.id
    )
    db.add(new_post)
    await db.commit()
@@ -49,30 +58,36 @@ async def get_post(post_id: int, db: Annotated[AsyncSession, Depends(get_db)]):
   return post
 
 @router.put("/{post_id}", response_model=PostResponse)
-async def update_post_full(post_id: int, post_data: PostCreate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_full(post_id: int, post_data: PostCreate, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
   result = await db.execute(select(models.Post)
                             .options(selectinload(models.Post.author))
                             .where(models.Post.id == post_id))
   post = result.scalars().first()
   if not post:
     raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Post not found")
-  if post_data.user_id != post.user_id:
-    result = await db.execute(select(models.User).where(models.User.id == post_data.user_id))
-    user = result.scalars().first()
-    if not user:
-      raise HTTPException(
-        status_code=status.HTTP_400_BAD_REQUEST,
-        detail="USer not found"
+  # if post_data.user_id != post.user_id:
+  #   result = await db.execute(select(models.User).where(models.User.id == post_data.user_id))
+  #   user = result.scalars().first()
+  #   if not user:
+  #     raise HTTPException(
+  #       status_code=status.HTTP_400_BAD_REQUEST,
+  #       detail="USer not found"
+  #     )
+  if post.user_id != current_user.id:
+    raise HTTPException(
+        status_code=status.HTTP_403_FORBIDDEN,
+        detail="Not authorized to update this post"
       )
   post.title = post_data.title
   post.content = post_data.content
-  post.user_id = post_data.user_id
+
   await db.commit()
   await db.refresh(post, attribute_names=["author"])
   return post
   
 @router.patch("/{post_id}", response_model=PostResponse)
-async def update_post_partial(post_id:int, post_data:PostUpdate, db: Annotated[AsyncSession, Depends(get_db)]):
+async def update_post_partial(post_id:int, post_data:PostUpdate,
+                               current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
   result = await db.execute(select(models.Post).where(models.Post.id == post_id))
   post = result.scalars().first()
   if not post:
@@ -80,6 +95,11 @@ async def update_post_partial(post_id:int, post_data:PostUpdate, db: Annotated[A
       status_code=status.HTTP_404_NOT_FOUND,
       detail="Post not found"
     )
+  if post.user_id != current_user.id:
+    raise HTTPException(
+        status_code=status.HTTP_403_FORBIDDEN,
+        detail="Not authorized to update this post"
+      )
   update_data = post_data.model_dump(exclude_unset=True) #exclude_unset
   for field, value in update_data.items():
     setattr(post, field, value)
@@ -88,7 +108,7 @@ async def update_post_partial(post_id:int, post_data:PostUpdate, db: Annotated[A
   return post
 
 @router.delete("/{post_id}", status_code=status.HTTP_204_NO_CONTENT)
-async def deleted_post(post_id:int, db: Annotated[AsyncSession, Depends(get_db)]):
+async def deleted_post(post_id:int, current_user: CurrentUser, db: Annotated[AsyncSession, Depends(get_db)]):
   result = await db.execute(select(models.Post).where(models.Post.id == post_id))
   post = result.scalars().first()
   if not post:
@@ -96,5 +116,10 @@ async def deleted_post(post_id:int, db: Annotated[AsyncSession, Depends(get_db)]
       status_code=status.HTTP_404_NOT_FOUND,
       detail="Post not found"
     )
+  if post.user_id != current_user.id:
+    raise HTTPException(
+        status_code=status.HTTP_403_FORBIDDEN,
+        detail="Not authorized to update this post"
+      )
   await db.delete(post)
   await db.commit()
